+++ SSToolkit/NSMutableArray+SSToolkitAdditions.m
@@ -6,7 +6,7 @@
 {
     if (self.count == 0) return;
     for (NSUInteger i = [self count] - 1; i > 0; i--) {
-        [self exchangeObjectAtIndex:arc4random_uniform(i + 1)
+        [self exchangeObjectAtIndex:arc4random_uniform((u_int32_t)i + 1)
                   withObjectAtIndex:i];
     }
 }
